Added login

+++ MvcLibrary/Controllers/BookController.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Http;
 using MvcLibrary.Data;
 using MvcLibrary.Models;
 
@@ -22,25 +23,20 @@ namespace MvcLibrary.Controllers
         // GET: Book
         public async Task<IActionResult> Index()
         {
-              return _context.Books != null ? 
-                          View(await _context.Books.ToListAsync()) :
-                          Problem("Entity set 'LibraryDbContext.Books'  is null.");
+            return _context.Books != null ?
+                        View(await _context.Books.ToListAsync()) :
+                        Problem("Entity set 'LibraryDbContext.Books'  is null.");
         }
 
         // GET: Book/Details/5
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null || _context.Books == null)
-            {
                 return NotFound();
-            }
 
-            var book = await _context.Books
-                .FirstOrDefaultAsync(m => m.Id == id);
+            var book = await _context.Books.FirstOrDefaultAsync(m => m.Id == id);
             if (book == null)
-            {
                 return NotFound();
-            }
 
             return View(book);
         }
@@ -48,16 +44,17 @@ namespace MvcLibrary.Controllers
         // GET: Book/Create
         public IActionResult Create()
         {
+            if (!IsAdmin()) return RedirectToAction("Index", "Login");
             return View();
         }
 
         // POST: Book/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Title,Author,Description")] Book book)
         {
+            if (!IsAdmin()) return RedirectToAction("Index", "Login");
+
             if (ModelState.IsValid)
             {
                 _context.Add(book);
@@ -70,30 +67,27 @@ namespace MvcLibrary.Controllers
         // GET: Book/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (!IsAdmin()) return RedirectToAction("Index", "Login");
+
             if (id == null || _context.Books == null)
-            {
                 return NotFound();
-            }
 
             var book = await _context.Books.FindAsync(id);
             if (book == null)
-            {
                 return NotFound();
-            }
+
             return View(book);
         }
 
         // POST: Book/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Title,Author,Description")] Book book)
         {
+            if (!IsAdmin()) return RedirectToAction("Index", "Login");
+
             if (id != book.Id)
-            {
                 return NotFound();
-            }
 
             if (ModelState.IsValid)
             {
@@ -105,13 +99,9 @@ namespace MvcLibrary.Controllers
                 catch (DbUpdateConcurrencyException)
                 {
                     if (!BookExists(book.Id))
-                    {
                         return NotFound();
-                    }
                     else
-                    {
                         throw;
-                    }
                 }
                 return RedirectToAction(nameof(Index));
             }
@@ -121,17 +111,14 @@ namespace MvcLibrary.Controllers
         // GET: Book/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null || _context.Books == null)
-            {
-                return NotFound();
-            }
+            if (!IsAdmin()) return RedirectToAction("Index", "Login");
 
-            var book = await _context.Books
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (book == null)
-            {
+            if (id == null || _context.Books == null)
                 return NotFound();
-            }
+
+            var book = await _context.Books.FirstOrDefaultAsync(m => m.Id == id);
+            if (book == null)
+                return NotFound();
 
             return View(book);
         }
@@ -141,23 +128,27 @@ namespace MvcLibrary.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!IsAdmin()) return RedirectToAction("Index", "Login");
+
             if (_context.Books == null)
-            {
                 return Problem("Entity set 'LibraryDbContext.Books'  is null.");
-            }
+
             var book = await _context.Books.FindAsync(id);
             if (book != null)
-            {
                 _context.Books.Remove(book);
-            }
-            
+
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool BookExists(int id)
         {
-          return (_context.Books?.Any(e => e.Id == id)).GetValueOrDefault();
+            return (_context.Books?.Any(e => e.Id == id)).GetValueOrDefault();
+        }
+
+        private bool IsAdmin()
+        {
+            return HttpContext.Session.GetString("IsAdmin") == "True";
         }
     }
 }
